# Fix: HttpClientFactory: resolving a typed client is significantly slower that resolving a named client.