--- a/src/HttpClientFactory/Http/perf/Properties/AssemblyInfo.cs
+++ b/src/HttpClientFactory/Http/perf/Properties/AssemblyInfo.cs
@@ -1,0 +1,6 @@
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using BenchmarkDotNet.Attributes;
+
+[assembly: AspNetCoreBenchmark]