See commits for details.


``` ini

BenchmarkDotNet=v0.10.13, OS=Windows 10.0.18362
Intel Core i7-8650U CPU 1.90GHz (Kaby Lake R), 1 CPU, 8 logical cores and 4 physical cores
.NET Core SDK=3.0.100-preview6-012264
  [Host]     : .NET Core 3.0.0-preview8-27919-01 (CoreCLR 4.700.19.36901, CoreFX 4.700.19.36803), 64bit RyuJIT
  Job-PYIOAJ : .NET Core 3.0.0-preview8-27919-01 (CoreCLR 4.700.19.36901, CoreFX 4.700.19.36803), 64bit RyuJIT

Runtime=Core  Server=True  Toolchain=.NET Core 3.0  
RunStrategy=Throughput  

```

**Before:**

|                                     Method |      Mean |      Error |     StdDev |    Median |     Op/s | Scaled | ScaledSD |  Gen 0 | Allocated |
|------------------------------------------- |----------:|-----------:|-----------:|----------:|---------:|-------:|---------:|-------:|----------:|
| &#39;use IHttpClientFactory with named client&#39; |  13.87 us |  0.2708 us |  0.3521 us |  13.84 us | 72,116.5 |   1.23 |     0.08 | 7.8125 |   3.18 KB |
| &#39;use IHttpClientFactory with typed client&#39; | 500.04 us | 17.8916 us | 52.1905 us | 515.59 us |  1,999.9 |  44.26 |     5.34 |      - |   8.63 KB |
|                           &#39;new HttpClient&#39; |  11.34 us |  0.2386 us |  0.6922 us |  11.28 us | 88,183.9 |   1.00 |     0.00 | 7.8125 |   2.66 KB |


**After:**

|                                     Method |      Mean |     Error |    StdDev |      Op/s | Scaled | ScaledSD |  Gen 0 | Allocated |
|------------------------------------------- |----------:|----------:|----------:|----------:|-------:|---------:|-------:|----------:|
| &#39;use IHttpClientFactory with named client&#39; | 11.065 us | 0.2203 us | 0.4549 us |  90,376.9 |   1.88 |     0.09 | 7.8125 |   3.17 KB |
| &#39;use IHttpClientFactory with typed client&#39; | 13.362 us | 0.2660 us | 0.6059 us |  74,839.8 |   2.27 |     0.11 | 7.8125 |   3.28 KB |
|                           &#39;new HttpClient&#39; |  5.877 us | 0.1116 us | 0.1194 us | 170,146.8 |   1.00 |     0.00 | 6.8359 |   2.67 KB |